Refactor #349 - make navigation compose
