+++ app/build.gradle.kts
@@ -73,6 +73,7 @@ dependencies {
     implementation(projects.core.android)
     implementation(projects.core.androidx.fragment)
     implementation(projects.core.androidx.lifecycle)
+    implementation(projects.core.androidx.navigationCompose)
     implementation(projects.core.androidx.recyclerview)
     implementation(projects.core.androidx.transition)
     implementation(projects.core.annotation)
