+++ core/androidx/navigation-compose/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+    alias(libs.plugins.nlab.android.library)
+    alias(libs.plugins.nlab.android.library.compose)
+}
+
+android {
+    namespace = "com.nlab.reminder.core.androidx.navigation.compose"
+}
+
+dependencies {
+    api(libs.androidx.navigation.compose)
+}
